+++ ui/DesignSystem/Component/Header/Large.svelte
@@ -1,13 +1,8 @@
 <script>
   import { createEventDispatcher } from "svelte";
 
-  import Avatar from "../Avatar.svelte";
-  import Button from "../Button.svelte";
-  import Icon from "../Icon";
-  import Title from "../Title.svelte";
-  import Text from "../Text.svelte";
-  import Numeric from "../Numeric.svelte";
-  import { Copyable } from "../../Component";
+  import { Avatar, Button, Icon, Title, Text, Numeric } from "../../Primitive";
+  import Copyable from "../Copyable.svelte";
 
   const dispatch = createEventDispatcher();
